# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Todostock API",
        "description": "Documentación oficial de los endpoints de la API de Todostock para Todotek S.A. Incluye autenticación por Sanctum, gestión de inventarios y facturación transaccional.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "Servidor Local de Desarrollo (Docker Sail)"
        }
    ],
    "paths": {
...
        "/products": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Obtener listado de productos con imágenes",
                "description": "Retorna una lista paginada de los productos con sus respectivas imágenes asociadas. Permite buscar de forma dinámica por nombre o SKU.",
                "operationId": "getProductsList",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Término de búsqueda para filtrar por nombre o SKU",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número de página para la paginación del servidor",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado de productos obtenido exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Teclado Mecánico RGB"
                                                    },
                                                    "sku": {
                                                        "type": "string",
                                                        "example": "TEC-MECH-01"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "Teclado con switches mecánicos y retroiluminación."
                                                    },
                                                    "stock": {
                                                        "type": "integer",
                                                        "example": 25
                                                    },
                                                    "price": {
                                                        "type": "string",
                                                        "example": "45.00"
                                                    },
                                                    "tax_percentage": {
                                                        "type": "string",
                                                        "example": "15.00"
                                                    },
                                                    "images": {
                                                        "type": "array",
                                                        "items": {
                                                            "properties": {
                                                                "id": {
                                                                    "type": "integer",
                                                                    "example": 1
                                                                },
                                                                "product_id": {
                                                                    "type": "integer",
                                                                    "example": 1
                                                                },
                                                                "file_path": {
                                                                    "type": "string",
                                                                    "example": "/storage/products/abc123image.png"
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "per_page": {
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 50
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Productos"
                ],
                "summary": "Registrar un nuevo producto con imágenes opcionales",
                "description": "Crea un producto en el catálogo. Al manejar subida de archivos físicos, requiere que la petición sea de tipo multipart/form-data.",
                "operationId": "storeProduct",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "sku",
                                    "stock",
                                    "price",
                                    "tax_percentage"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Mouse Gamer Óptico"
                                    },
                                    "sku": {
                                        "type": "string",
                                        "example": "MOU-GAME-02"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Mouse ergonómico con sensor de 16000 DPI."
                                    },
                                    "stock": {
                                        "type": "integer",
                                        "example": 15
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 29.99
                                    },
                                    "tax_percentage": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 15
                                    },
                                    "images[]": {
                                        "description": "Arreglo de imágenes físicas a subir (jpeg, png, webp. Máx 2MB por archivo)",
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Producto creado exitosamente con sus imágenes vinculadas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 6
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Mouse Gamer Óptico"
                                        },
                                        "sku": {
                                            "type": "string",
                                            "example": "MOU-GAME-02"
                                        },
                                        "stock": {
                                            "type": "integer",
                                            "example": 15
                                        },
                                        "price": {
                                            "type": "number",
                                            "example": 29.99
                                        },
                                        "tax_percentage": {
                                            "type": "number",
                                            "example": 15
                                        },
                                        "images": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación (Ej: SKU ya registrado o formato de imagen no soportado)"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Introduce el token 'access_token' recibido en el login con el formato: Bearer {token}",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}